pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "JoyIT_WLGamepad"
version = "1.0.0"
dependencies = ['RPi.GPIO>=0.7.0']
authors = [
  { name="Joy-IT", email="service@joy-it.net" },
]
description = "Library to use the SBC-WLGamepad with the Raspberry Pi"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/joy-it/JoyIT_WLGamepad"
Issues = "https://github.com/joy-it/JoyIT_WLGamepad/issues"